rename: Attendance studyEntity로 이동
